パーミッションチェックを行うためのスプラッシュを用意
